+++ python_3/solutions/soln_kaggle_covid19.py
@@ -1,0 +1,166 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Jul  3 17:51:11 2020
+
+@author: Ashish
+"""
+# reference: https://www.kaggle.com/tarunkr/covid-19-case-study-analysis-viz-comparisons?select=covid_19_india.csv#Introduction-to-COVID-19
+
+import pandas as pd
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib import ticker 
+import pycountry_convert as pc
+import folium
+import branca
+from datetime import datetime, timedelta,date
+from scipy.interpolate import make_interp_spline, BSpline
+import plotly.express as px
+import json, requests
+import calmap
+
+from keras.layers import Input, Dense, Activation, LeakyReLU, Dropout
+from keras import models
+from keras.optimizers import RMSprop, Adam
+
+import warnings
+warnings.filterwarnings('ignore')
+
+#%matplotlib inline
+
+# load the data
+df_confirmed = pd.read_csv("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv")
+print(df_confirmed.head())
+df_deaths = pd.read_csv("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv")
+print(df_deaths.head())
+df_recovrd = pd.read_csv("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_recovered_global.csv")
+
+# data preprocessing
+# replace blanks with NAN
+df_confirmed = df_confirmed.replace(np.nan,'',regex=True)
+df_deaths = df_deaths.replace(np.nan,'',regex=True)
+df_recovrd = df_recovrd.replace(np.nan,'',regex=True)
+
+# defining custom function
+def plot_params(ax,axis_label= None, plt_title = None,label_size=15, axis_fsize = 15, title_fsize = 20, scale = 'linear' ):
+    # Tick-Parameters
+    ax.xaxis.set_minor_locator(ticker.AutoMinorLocator())
+    ax.yaxis.set_minor_locator(ticker.AutoMinorLocator())
+    ax.tick_params(which='both', width=1,labelsize=label_size)
+    ax.tick_params(which='major', length=6)
+    ax.tick_params(which='minor', length=3, color='0.8')
+    
+    # Grid
+    plt.grid(lw = 1, ls = '-', c = "0.7", which = 'major')
+    plt.grid(lw = 1, ls = '-', c = "0.9", which = 'minor')
+
+    # Plot Title
+    plt.title( plt_title,{'fontsize':title_fsize})
+    
+    # Yaxis sacle
+    plt.yscale(scale)
+    plt.minorticks_on()
+    # Plot Axes Labels
+    xl = plt.xlabel(axis_label[0],fontsize = axis_fsize)
+    yl = plt.ylabel(axis_label[1],fontsize = axis_fsize)
+    
+def visualize_covid_cases(confirmed, deaths, continent=None , country = None , state = None, period = None, figure = None, scale = "linear"):
+    x = 0
+    if figure == None:
+        f = plt.figure(figsize=(10,10))
+        # Sub plot
+        ax = f.add_subplot(111)
+    else :
+        f = figure[0]
+        # Sub plot
+        ax = f.add_subplot(figure[1],figure[2],figure[3])
+    ax.set_axisbelow(True)
+    plt.tight_layout(pad=10, w_pad=5, h_pad=5)
+    
+    stats = [confirmed, deaths]
+    label = ["Confirmed", "Deaths"]
+    
+    if continent != None:
+        params = ["continent",continent]
+    elif country != None:
+        params = ["country",country]
+    else: 
+        params = ["All", "All"]
+    color = ["darkcyan","crimson"]
+    marker_style = dict(linewidth=3, linestyle='-', marker='o',markersize=4, markerfacecolor='#ffffff')
+    for i,stat in enumerate(stats):
+        if params[1] == "All" :
+            cases = np.sum(np.asarray(stat.iloc[:,5:]),axis = 0)[x:]
+        else :
+            cases = np.sum(np.asarray(stat[stat[params[0]] == params[1]].iloc[:,5:]),axis = 0)[x:]
+        date = np.arange(1,cases.shape[0]+1)[x:]
+        plt.plot(date,cases,label = label[i]+" (Total : "+str(cases[-1])+")",color=color[i],**marker_style)
+        plt.fill_between(date,cases,color=color[i],alpha=0.3)
+
+    if params[1] == "All" :
+        Total_confirmed = np.sum(np.asarray(stats[0].iloc[:,5:]),axis = 0)[x:]
+        Total_deaths = np.sum(np.asarray(stats[1].iloc[:,5:]),axis = 0)[x:]
+    else :
+        Total_confirmed =  np.sum(np.asarray(stats[0][stat[params[0]] == params[1]].iloc[:,5:]),axis = 0)[x:]
+        Total_deaths = np.sum(np.asarray(stats[1][stat[params[0]] == params[1]].iloc[:,5:]),axis = 0)[x:]
+        
+    text = "From "+stats[0].columns[5]+" to "+stats[0].columns[-1]+"\n"
+    text += "Mortality rate : "+ str(int(Total_deaths[-1]/(Total_confirmed[-1])*10000)/100)+"\n"
+    text += "Last 5 Days:\n"
+    text += "Confirmed : " + str(Total_confirmed[-1] - Total_confirmed[-6])+"\n"
+    text += "Deaths : " + str(Total_deaths[-1] - Total_deaths[-6])+"\n"
+    text += "Last 24 Hours:\n"
+    text += "Confirmed : " + str(Total_confirmed[-1] - Total_confirmed[-2])+"\n"
+    text += "Deaths : " + str(Total_deaths[-1] - Total_deaths[-2])+"\n"
+    
+    plt.text(0.02, 0.78, text, fontsize=15, horizontalalignment='left', verticalalignment='top', transform=ax.transAxes,bbox=dict(facecolor='white', alpha=0.4))
+    
+    # Plot Axes Labels
+    axis_label = ["Days ("+df_confirmed.columns[5]+" - "+df_confirmed.columns[-1]+")","No of Cases"]
+    
+    # Plot Parameters
+    plot_params(ax,axis_label,scale = scale)
+    
+    # Plot Title
+    if params[1] == "All" :
+        plt.title("COVID-19 Cases World",{'fontsize':25})
+    else:   
+        plt.title("COVID-19: "+params[1] ,{'fontsize':25})
+        
+    # Legend Location
+    l = plt.legend(loc= "best",fontsize = 15)
+    
+    if figure == None:
+        plt.show()
+        
+def get_total_cases(cases, country = "All"):
+    if(country == "All") :
+        return np.sum(np.asarray(cases.iloc[:,5:]),axis = 0)[-1]
+    else :
+        return np.sum(np.asarray(cases[cases["country"] == country].iloc[:,5:]),axis = 0)[-1]
+    
+def get_mortality_rate(confirmed,deaths, continent = None, country = None):
+    if continent != None:
+        params = ["continent",continent]
+    elif country != None:
+        params = ["country",country]
+    else :
+        params = ["All", "All"]
+    
+    if params[1] == "All" :
+        Total_confirmed = np.sum(np.asarray(confirmed.iloc[:,5:]),axis = 0)
+        Total_deaths = np.sum(np.asarray(deaths.iloc[:,5:]),axis = 0)
+        mortality_rate = np.round((Total_deaths/(Total_confirmed+1.01))*100,2)
+    else :
+        Total_confirmed =  np.sum(np.asarray(confirmed[confirmed[params[0]] == params[1]].iloc[:,5:]),axis = 0)
+        Total_deaths = np.sum(np.asarray(deaths[deaths[params[0]] == params[1]].iloc[:,5:]),axis = 0)
+        mortality_rate = np.round((Total_deaths/(Total_confirmed+1.01))*100,2)
+    
+    return np.nan_to_num(mortality_rate)
+def dd(date1,date2):
+    return (datetime.strptime(date1,'%m/%d/%y') - datetime.strptime(date2,'%m/%d/%y')).days
+
+
+out = ""#+"output/"
+
+
